首页翻译修改以及相机调整
