+++ app/build.gradle
@@ -25,6 +25,7 @@ android {
             includeAndroidResources = true
         }
     }
+    buildToolsVersion '26.0.2'
 }
 
 repositories {
@@ -32,30 +33,23 @@ repositories {
 }
 
 dependencies {
-    implementation fileTree(dir: 'libs', include: ['*.jar'])
+    implementation fileTree(include: ['*.jar'], dir: 'libs')
     implementation "org.jetbrains.kotlin:kotlin-stdlib-jre7:$kotlin_version"
     implementation 'com.android.support:appcompat-v7:26.1.0'
     implementation 'com.android.support.constraint:constraint-layout:1.0.2'
     compile 'com.android.support:design:26.1.0'
-
-
     compile 'com.squareup.okhttp3:okhttp:3.10.0'
     compile 'com.squareup.okio:okio:1.14.0'
     compile 'com.google.code.gson:gson:2.8.2'
-
     compile 'com.android.support:cardview-v7:26.1.0'
     implementation 'de.hdodenhof:circleimageview:2.2.0'
-
     compile 'com.github.everhad:AndroidRatingStar:v1.0.1'
-
     compile 'com.jaren:likeview:1.1.0'
-
     testImplementation 'junit:junit:4.12'
     testCompile 'org.mockito:mockito-core:1.10.19'
     testCompile 'org.powermock:powermock-module-junit4:1.6.2'
     testCompile 'org.powermock:powermock-module-junit4-rule:1.6.2'
     testCompile 'org.powermock:powermock-api-mockito:1.6.2'
-
     androidTestImplementation 'com.android.support.test:runner:1.0.1'
     androidTestImplementation 'com.android.support.test.espresso:espresso-core:3.0.1'
 }
