+++ app/src/main/res/layout/activity_main.xml
@@ -7,15 +7,24 @@
     android:layout_height="match_parent"
     tools:context="com.lmmnb.guangzhouculturehelper.app.view.MainActivity">
 
+    <TextView
+        android:id="@+id/tv4translate"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_marginLeft="20dp"
+        android:layout_marginTop="140px"/>
+
     <com.lmmnb.guangzhouculturehelper.app.view.widget.MyEditText
         android:id="@+id/et1"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:layout_alignParentLeft="true"
         android:layout_alignParentStart="true"
-        android:layout_alignParentTop="true">
-
+        android:layout_alignParentTop="true"
+        android:imeOptions="actionSend"
+        android:inputType="textMultiLine">
     </com.lmmnb.guangzhouculturehelper.app.view.widget.MyEditText>
+
 
     <ImageButton
         android:id="@+id/clearBtn"
