+++ app/src/main/res/layout/myet.xml
@@ -14,7 +14,7 @@
         android:gravity="start|left"
         android:hint="在此输入要翻译的文本..."
         android:paddingLeft="20dp"
-        android:paddingTop="20dp"
+        android:paddingTop="40px"
         android:imeOptions="actionDone"
         android:singleLine="true"
         />
