--- app/src/main/res/layout/picture_food.xml
@@ -1,52 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<RelativeLayout
-    xmlns:android="http://schemas.android.com/apk/res/android" android:layout_width="match_parent"
-    android:layout_height="match_parent">
-
-    <ImageView
-        android:id="@+id/picture"
-        android:src="@drawable/camera_backgraound"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent" />
-
-
-
-    <RelativeLayout
-        android:id="@+id/relative_layout"
-        android:layout_width="match_parent"
-        android:layout_height="200dp"
-        android:layout_alignParentBottom="true"
-        android:background="#CC5c6bc0">
-
-        <TextView
-            android:text="对准食物拍照"
-            android:textColor="#E4EFFD"
-            android:textSize="18sp"
-            android:layout_centerHorizontal="true"
-            android:layout_width="wrap_content"
-            android:layout_marginTop="12dp"
-            android:layout_height="wrap_content"
-            android:id="@+id/textView" />
-
-        <Button
-            android:id="@+id/camear_return"
-            android:layout_width="50dp"
-            android:layout_height="50dp"
-            android:layout_centerVertical="true"
-            android:layout_marginEnd="35dp"
-            android:layout_marginRight="35dp"
-            android:layout_toStartOf="@+id/textView"
-            android:layout_toLeftOf="@+id/textView"
-            android:background="@drawable/ic_close_black_24dp" />
-
-        <ImageButton
-            android:id="@+id/camera"
-            android:background="@drawable/camera_btn_background"
-            android:src="@drawable/ic_camera_alt_black_24dp"
-            android:layout_centerInParent="true"
-            android:layout_width="90dp"
-            android:layout_height="90dp"
-            />
-    </RelativeLayout>
-
-</RelativeLayout>
