+++ app/src/main/res/xml/file_paths.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
 <paths
     xmlns:android="http://schemas.android.com/apk/res/android">
-    <external-path name="my_images" path="" />
+    <external-path name="my_images" path="Android/data/com.lmmnb.guangzhouculturehelper/files/Pictures" />
 </paths>
